+++ src/test/unit/ntypes/SdrMetricsTest.cpp
@@ -67,7 +67,7 @@ TEST(SdrMetrics, TestSparsityShortTerm) {
     SDR_Proxy B( A );
     Real period = 10u;
     Real alpha  = 1.0f / period;
-    SDR_Sparsity S( B, period );
+    SDR_Sparsity S( B, (UInt)period );
 
     A.setDense(SDR_dense_t{ 1 });
     ASSERT_FLOAT_EQ( S.sparsity, 1.0f );
@@ -250,7 +250,7 @@ TEST(SdrMetrics, TestAF_LongTerm) {
     SDR_Proxy B( A );
     SDR_ActivationFrequency F( B, period );
 
-    vector<Real> test_sparsity{ 0.0f, 0.02f, 0.05, 1.0f, 0.25f, 0.5f };
+    vector<Real> test_sparsity{ 0.0f, 0.02f, 0.05f, 1.0f, 0.25f, 0.5f };
 
     for(const auto &sparsity : test_sparsity) {
         for(UInt i = 0; i < runtime; i++)
@@ -378,10 +378,10 @@ TEST(SdrMetrics, TestOverlap_Example) {
     SDR A({ 10000u });
     SDR_Proxy Px( A );
     SDR_Overlap B( Px, 1000u );
-    A.randomize( 0.05 );
-    A.addNoise( 0.95 );         //   5% overlap
-    A.addNoise( 0.55 );         //  45% overlap
-    A.addNoise( 0.72 );         //  28% overlap
+    A.randomize( 0.05f );
+    A.addNoise( 0.95f );         //   5% overlap
+    A.addNoise( 0.55f );         //  45% overlap
+    A.addNoise( 0.72f );         //  28% overlap
     ASSERT_EQ( B.overlap,  0.28f );
     ASSERT_EQ( B.min(),    0.05f );
     ASSERT_EQ( B.max(),    0.45f );
